修改路径错误和镜像名称截取错误

+++ start_bootstrapper_container.sh
@@ -2,25 +2,43 @@
 
 # start_bootstrapper_container.sh load docker image from bsroot and
 # then push them to registry
+if [[ "$#" -gt 1 ]]; then
+    echo "Usage: start_bootstrapper_contaienr.sh [bsroot-path]"
+    exit 1
+elif [[ "$#" -ne 1 ]]; then
+    BSROOT=/bsroot
+else
+    BSROOT=$1
+fi
+
+if [[ ! -d $BSROOT ]]; then
+    echo "$BSROOT is not a directory"
+    exit 2
+fi
+
+if [[ $BSROOT != /* ]]; then
+  echo "bsroot path not start with / !"
+  exit 1
+fi
 
 # push k8s images to registry from bsroot
-BOOTATRAPPER_DOMAIN=`grep "dockerdomain:" /bsroot/config/cluster-desc.yml | awk '{print $2}' | sed 's/"//g' | sed 's/ //g'`
-HYPERKUBE_VERSION=`grep "hyperkube_version:" /bsroot/config/cluster-desc.yml | awk '{print $2}' | sed 's/ //g' | sed -e 's/^"//' -e 's/"$//'`
-PAUSE_VERSION=`grep "pause_version:" /bsroot/config/cluster-desc.yml | awk '{print $2}' | sed 's/ //g' | sed -e 's/^"//' -e 's/"$//'`
-FLANNEL_VERSION=`grep "flannel_version:" /bsroot/config/cluster-desc.yml | awk '{print $2}' | sed 's/ //g' | sed -e 's/^"//' -e 's/"$//'`
+BOOTATRAPPER_DOMAIN=`grep "dockerdomain:" $BSROOT/config/cluster-desc.yml | awk '{print $2}' | sed 's/"//g' | sed 's/ //g'`
+HYPERKUBE_VERSION=`grep "hyperkube_version:" $BSROOT/config/cluster-desc.yml | awk '{print $2}' | sed 's/ //g' | sed -e 's/^"//' -e 's/"$//'`
+PAUSE_VERSION=`grep "pause_version:" $BSROOT/config/cluster-desc.yml | awk '{print $2}' | sed 's/ //g' | sed -e 's/^"//' -e 's/"$//'`
+FLANNEL_VERSION=`grep "flannel_version:" $BSROOT/config/cluster-desc.yml | awk '{print $2}' | sed 's/ //g' | sed -e 's/^"//' -e 's/"$//'`
 
 # Config Registry tls
 mkdir -p /etc/docker/certs.d/bootstrapper:5000
 rm -rf /etc/docker/certs.d/bootstrapper:5000/*
-cp bsroot/tls/ca.pem /etc/docker/certs.d/bootstrapper:5000/ca.crt
+cp $BSROOT/tls/ca.pem /etc/docker/certs.d/bootstrapper:5000/ca.crt
 
 echo "127.0.0.1 bootstrapper" >> /etc/hosts
 
-docker load < /bsroot/bootstrapper.tar > /dev/null 2>&1 || { echo "Docker can not load bootstrapper.tar!"; exit 1; }
+docker load < $BSROOT/bootstrapper.tar > /dev/null 2>&1 || { echo "Docker can not load bootstrapper.tar!"; exit 1; }
 docker run -d --net=host \
   --privileged \
   -v /var/run/docker.sock:/var/run/docker.sock \
-  -v /bsroot:/bsroot \
+  -v $BSROOT:/bsroot \
   bootstrapper
 
 # Sleep 3 seconds, waitting for registry started.
@@ -37,7 +55,7 @@ DOCKER_IMAGES=("typhoon1986/hyperkube-amd64:${HYPERKUBE_VERSION}" \
 len=${#DOCKER_IMAGES[@]}
 for ((i=0;i<len;i++)); do
   DOCKER_IMAGE=${DOCKER_IMAGES[i]}
-  DOCKER_TAR_FILE=`echo /bsroot/${DOCKER_IMAGE}.tar | sed "s/:/_/g" |awk -F'/' '{print $2}'`
+  DOCKER_TAR_FILE=$BSROOT/$(echo ${DOCKER_IMAGE}.tar | sed "s/:/_/g" |awk -F'/' '{print $2}')
   DOCKER_TAG_NAME=`echo $BOOTATRAPPER_DOMAIN:5000/${DOCKER_IMAGE} | awk -F'/' '{print $1"/"$3}'`
   docker load < $DOCKER_TAR_FILE
   docker tag $DOCKER_IMAGE $DOCKER_TAG_NAME
